Использование собственного исключения

+++ opinions_app/api_views.py
@@ -3,13 +3,16 @@ from random import randrange
 from flask import jsonify, request
 
 from . import app, db
+from .error_handlers import InvalidAPIUsage
 from .models import Opinion
 from .views import random_opinion
 
 
 @app.route('/api/opinions/<int:id>/', methods=['GET'])
 def get_opinion(id):
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     return jsonify({'opinion': opinion.to_dict()}), 200
 
 
@@ -17,9 +20,10 @@ def get_opinion(id):
 def update_opinion(id):
     data = request.get_json()
     if 'text' in data and Opinion.query.filter_by(text=data['text']).first() is not None:
-        return jsonify({'error': 'Такое мнение уже есть в базе данных'}), 400
-    opinion = Opinion.query.get_or_404(id)
-
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных')
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     opinion.title = data.get('title', opinion.title)
     opinion.text = data.get('text', opinion.text)
     opinion.source = data.get('source', opinion.source)
@@ -31,7 +35,9 @@ def update_opinion(id):
 
 @app.route('/api/opinions/<int:id>/', methods=['DELETE'])
 def delete_opinion(id):
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     db.session.delete(opinion)
     db.session.commit()
     return '', 204
@@ -48,9 +54,9 @@ def get_opinions():
 def add_opinion():
     data = request.get_json()
     if 'title' not in data or 'text' not in data:
-        return jsonify({'error': 'В запросе отсутствуют обязательные поля'}), 400
+        raise InvalidAPIUsage('В запросе отсутствуют обязательные поля')
     if Opinion.query.filter_by(text=data['text']).first() is not None:
-        return jsonify({'error': 'Такое мнение уже есть в базе данных'}), 400
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных')
     opinion = Opinion()
     opinion.from_dict(data)
 
@@ -62,4 +68,6 @@ def add_opinion():
 @app.route('/api/get-random-opinion/', methods=['GET'])
 def get_random_opinion():
     opinion = random_opinion()
-    return jsonify({'opinion': opinion.to_dict()}), 200
+    if opinion is not None:
+        return jsonify({'opinion': opinion.to_dict()}), 200
+    raise InvalidAPIUsage('В базе данных нет мнений', 404)
